Restructure EKS directory
