+++ infra/eks/policies/allow_privilege_escalation.rego
