+++ infra/eks/policies/privileged.rego
